requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
djangorestframework==3.10.3
mysqlclient==1.4.6
Pillow==6.2.1
pip==19.3.1
PyMySQL==0.9.3
pytz==2019.3
setuptools==42.0.1
sqlparse==0.3.0
wheel==0.33.6